requirements.txt
Flask==2.2.2  # Web framework for building the application
dotenv==0.20.0  # Loads environment variables from a .env file
google-api-python-client==2.49.0  # Client library for accessing Google APIs
google-generativeai==0.1.1  # Library for interacting with Google's generative AI models
PyPDF2==1.31.0  # Library for reading and extracting text from PDF files
langchain==0.15.1  # Library for building text-based AI pipelines
langchain-google-genai==0.15.1  # Adapter for using Google generative AI models with Langchain
FAISS-cpu==1.8.1  # Library for efficient similarity search
requests==2.28.1  # Library for making HTTP requests